Move ErrorIngestionCustomCheck to a separate file.
